==> notebooks/05_train_test_split.ipynb <==
# 05 - Train-Test-Split

import pandas as pd
from sklearn.model_selection import train_test_split

# Display all rows and columns
pd.set_option('display.max_colwidth', None)  # Show full content in each cell
pd.set_option('display.max_rows', None)      # Show all rows
pd.set_option('display.max_columns', None)   # Show all columns

df = pd.read_parquet("../data/processed/hr_data_encoded.parquet")

df.head()

# Drop Employee_Name
df = df.drop(columns=["Employee_Name"], errors="ignore")

# Define features and target
X = df.drop(columns=["GoodFit"])
y = df["GoodFit"]

# Split the data into training and testing subsets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42, stratify=y)

output_path = "../data/processed/"

X_train.to_parquet(f"{output_path}X_train.parquet")
X_test.to_parquet(f"{output_path}X_test.parquet")
y_train.to_frame().to_parquet(f"{output_path}y_train.parquet")
y_test.to_frame().to_parquet(f"{output_path}y_test.parquet")
